Machine_Learning-Image_Denoising/requirements.txt
waitress
flask-cors
flask
tensorflow-cpu